Post-processing Turrialba-Masatepe yield simulations

+++ BC_CAF2021_Turrialba_Masatepe.R
@@ -4,7 +4,7 @@
    rm(list = ls())
 
 ## 1. INITIALISE MCMC ##
-   nChain <- as.integer(2e5)
+   nChain <- as.integer(7e4)
    source('BC/BC_CAF2021_MCMC_init_Turrialba_Masatepe.R')
 
 ## 2. RUNNING THE MCMC ##
